Rename cosmos namespace and remove the word "bank integration" from it
