+++ Spbs.Ui/Data/Cosmos/CosmosDocument.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json;
 
-namespace Spbs.Ui.Data.Cosmos;
+namespace Spbs.Data.Cosmos;
 
 public interface ICosmosData
 {
