+++ Spbs.Ui/Features/RecurringExpenses/Repositories/RecurringExpenseReader.cs
@@ -8,7 +8,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Shared.Utilities;
 using Spbs.Shared.Data;
-using Spbs.Ui.Data.Cosmos;
+using Spbs.Data.Cosmos;
 using Spbs.Ui.Features.BankIntegration;
 
 namespace Spbs.Ui.Features.RecurringExpenses;
