+++ Spbs.Ui/Features/Visualization/DataAccess/ExpenseBatchReader.cs
@@ -8,7 +8,7 @@ using Microsoft.Azure.Cosmos.Linq;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Spbs.Shared.Data;
-using Spbs.Ui.Data.Cosmos;
+using Spbs.Data.Cosmos;
 using Spbs.Ui.Features.BankIntegration;
 using Spbs.Ui.Features.Expenses;
 using Spbs.Ui.Features.Visualization.Models;
